Add files via upload
This is the Excel macro that was used to sort the data into a usable form. It creates two sheets, one with the values sorted in x and y columns for all of the readings, and the other with all of the depths that the readings occurred at.
It also contains the MATLAB file which was used to test the algorithm. The particular file was the version used to determine what values each of the parameters should be in order to minimise the error.